refactor: move NotFound component to ui/ directory
- Move NotFound component from app/components/ to app/components/ui/
- Update import paths in both 404 pages
